requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
dashscope>=1.14.0
matplotlib>=3.7.0
numpy>=1.24.0

